from error solved

+++ webapp/forms.py
@@ -159,10 +159,10 @@ class ContactForm(forms.Form):
 class SearchForm(forms.Form):
     plaats_postcode = forms.CharField(widget=forms.TextInput(attrs={'placeholder': 'Zoek op plaats of postcode'}))
     pand_type = forms. NullBooleanField(widget=forms.CheckboxInput())
-    aantal_slaapkamers = forms.BigIntegerField(widget=forms.TextInput(attrs={'id': 'aantal-slaapkamers', 'readonly style': 'border:0; color:#f6931f; font-weight:bold;'}))
-    aantal_badkamers = forms.BigIntegerField(widget=forms.TextInput(attrs={'id': 'aantal-badkamers', 'readonly style': 'border:0; color:#f6931f; font-weight:bold;'}))
-    aantal_verdiepen = forms.BigIntegerField(widget=forms.TextInput(attrs={'id': 'aantal-verdiepen', 'readonly style': 'border:0; color:#f6931f; font-weight:bold;'}))
-    prijs_range = forms.BigIntegerField(widget=forms.TextInput(attrs={'id': 'prijs-range', 'readonly style': 'border:0; color:#f6931f; font-weight:bold;'}))
+    aantal_slaapkamers = forms.IntegerField(widget=forms.TextInput(attrs={'id': 'aantal-slaapkamers', 'readonly style': 'border:0; color:#f6931f; font-weight:bold;'}))
+    aantal_badkamers = forms.IntegerField(widget=forms.TextInput(attrs={'id': 'aantal-badkamers', 'readonly style': 'border:0; color:#f6931f; font-weight:bold;'}))
+    aantal_verdiepen = forms.IntegerField(widget=forms.TextInput(attrs={'id': 'aantal-verdiepen', 'readonly style': 'border:0; color:#f6931f; font-weight:bold;'}))
+    prijs_range = forms.DecimalField(widget=forms.TextInput(attrs={'id': 'prijs-range', 'readonly style': 'border:0; color:#f6931f; font-weight:bold;'}))
 
     tags = forms.CharField(widget=forms.TextInput(attrs={'id': 'tags', 'data-role': 'tagsinput'}))
